1534: use container name as node name for puppet.
sets the puppet configuration to use a fact as the node name. this fact
is an executable (container.sh) that outputs the value of the
environment variable CONTAINER. this variable should be set in the
container yaml description to the name we want to expose to the puppet
master.

also adds a master auth.conf to allow nodes to provide an alternative
node name other than certname.

Fixes #OS-1534.