1492: add ceph container to the dev setup.
adds a ceph pod and container to the environment.

given that ceph is very picky regarding ip addresses, and these are
needed for the setup inside the container, we cannot rely on the usual
kubernetes service concept and need to setup the container with host
networking.

in addition we expose the docker0 ip so that all other containers can
contact the monitor and the osds later.

Fixes #OS-1492.