workaround for ldap setup (ldif schema in tmp).
need to figure out later how to best handle this, but for now the ldif
is stored /tmp and requires a manual mv later to the cn=schema dir in
openldap and a slapd restart.

reason is the how the osixia setup is done with their automated sh
scripts which prevent us from moving the file to cn=schema in the image
(the dir is recreated by the sh script).