make ldap container more flexible.
changed docker ldap container to a simpler one, which takes the base
ldif data from a volume instead of inside the image. this should make it
more flexible for future updates.

also added a patch on keystone to totally ignore tokenGroups so that we
don't need to update the ldap schema for now.