add volume, add wait condition for master/slave.
add a volume to be attached to the jenkins master (existing volume, uuid
passed as a param).

add a wait condition for the slaves to be launched only after the
jenkins service is running. like this we make sure the master is
reachable when the java swarm client runs on the slaves.