1819: adapt layout to get build from last.
move the controller launch out of 'restart' so that we can apply the
tag passed when launching the openstack containers.

rename 'rebuild' command to 'launch', taking a tag as an arg - by
default (no tag) it rebuilds from scratch, otherwise launches the
container from the tag and runs puppet after.

add a new 'last' command, equal to 'launch last'.

add a MySQL dockerfile, equal to upstream apart that the db data dir is
not a volume - we need to commit the data with the container to
reproduce errors later and to 'build from last'.

Fixes #1819.